plugins/eni: add ipv6 support
Extend interfaces and the engine to add support for adding an ipv6 address to
the ENI. The logic of determining the ipv6 address of the gateway is not
covered as a part of this. However, it still covers assigning the ipv6 address
to the interface and starting/stopping the dhclient process for the same.