plugins/ecs-bridge: add ecs-bridge plugin.
The ecs-bridge plugin creates a bridge and invokes the ipam
plugin if specified. It is very minimalistic in terms of
capabilities. The bridge name and the mtu size are the
only parameters that are supported for this plugin. It also
deletes all default routes for the interface added to
connect to the container's namespace