eni/engine: retry dhclient start & imds queries
Benchmarking experiments show that there's usually 3-6s delay in
IPv6 address path being populated in Instance Metadata for ENI.

We also need to take into account delays in cleaning up of the
dhclient processes running in the default network namespace.

This commit adds retries for both of these to account for these
delays. This will result in a latency increase in surfacing errors
when there are failures such as wrong MAC Address being specified
in the CNI config. However, we err on the side of correctness,
fault tolerance and caution, and deem that as an acceptable cost.