eni: ignore 'file exists' for default ipv6 route
Since the default ipv6 gateway can also be configured based on the
ipv6 address when the address is assigned, we sometimes get the
"file exists" error when adding the default route for the ipv6
gateway. This change ignores that error.

There are also logging improvements to log either the eni id or the
device name in various log messages for the eni plugin.