eni: add ipv4 subnet gateway from config
Added a new config field for supplying ipv4 subnet gateway address for
the caller. When that's specified the plugin can use that directly
instead of looking up that information from instance metadata service.

Also, got rid of the check where the plugin makes sure that the eni id
and the mac address match by looking up instance metadata. This leads to
addtional delays starting the plugin while adding very little value to
the functionality. Hence, the ENI plugin doesn't do that check anymore.
It's left to the caller instead should make sure that correct values
are passed to the plugin, which the ECS agent is currently doing anyway.