plugins/ecs-bridge/engine: fix race conditions.
When multiple awsvpc tasks launched at the same time, creating and configuring the bridge might fail because of race conditions. Fixing by:
(1) When creating the bridge, check whether the bridge exists after failing to create it.
(2) When assigning ip address to the bridge, check whether the ip address is already assigned after failing to assign it.