Release a new version of the plugins (#96)
* Adding security disclosure section to README file.

* updating go version used by Travis

* updating issues template with instruction on how to report security issues

* plugins/ecs-bridge/engine: fix race conditions.

When multiple awsvpc tasks launched at the same time, creating and configuring the bridge might fail because of race conditions. Fixing by:
(1) When creating the bridge, check whether the bridge exists after failing to create it.
(2) When assigning ip address to the bridge, check whether the ip address is already assigned after failing to assign it.

* update cni version to 2019.06.0

* eni plugin: adding 'stay-down' config

The `stay-down` config specifies if the interface should stay down. If
this is set, routes, addresses etc for the interface will not be set.

* eni plugin: don't bring up eni when 'stay-down' is set

Modified the eni plugins engine code to only move the ENI to the
target network namespace when 'stay-down' flag is set to 'true'.

* eni mocks: regenerate to capture `Engine` interface changes

Regenrate mocks as `Engine` interface has been modified to
accomodate the `stay-down` flag. Also updated the vendored
`gomock` package.

* eni plugin: 'stay-down' config tests

Added unit tests for the 'stay-down' ENI config.

* changelog entry for 'stay-down' config

* Add option to set interface MTU in eni plugin

* update the version to 2019.10.0