Always run DeleteVeth on cleanup
Previously to this change we would skip running DeleteVeth on cleanup if
cni plugins was not provided with an IPAM config to delete.

The agent codebase, however, does not always pass in an IPAM config on
purpose, when it is only cleaning up a single task, see
https://github.com/aws/amazon-ecs-agent/blame/71a0f6765cc469913a0b09728196feb0d0d33e37/agent/engine/docker_task_engine.go#L1779
and https://github.com/aws/amazon-ecs-agent/blame/71a0f6765cc469913a0b09728196feb0d0d33e37/agent/engine/docker_task_engine.go#L1760

With this change, when we skip the IPAM delete, we no longer immediately
return from the del function. Instead we let the code fall through to
the engine.DeleteVeth call to make sure that veth interfaces are always
cleaned up when a task is being cleaned up.