Set static MAC address when creating ecs-bridge bridge
The ecs-bridge bridge currently is created without setting a static MAC
address. This means that the bridge MAC address inherits the 'lowest'
MAC address of all of it's connected veth interfaces.

What this means for ECS tasks using awsvpc networking mode is that the
MAC address of ecs-bridge can be changing as tasks are killed an added
to an instance (on the ec2 launch type).

This presents a race condition where some tasks might have an arpcache
entry pointing to the old MAC address when a task is killed and the
ecs-bridge MAC changes. In rare cases this can lead to brief periods of
lost network connectivity within the local network on the host (such as
requests to the ECS Task Metadata Endpoint or the ECS credentials
endpoint).

By setting a static MAC address at creation we disable this behavior of
inheriting the 'lowest' MAC address. This means that the ecs-bridge MAC
stays constant through the entire lifetime of the ECS container
instance.